Edit wire1, wire2, breadboard, create computeMatrix
Edited the models for wire1, wire2, and breadboard. Breadboard's model
dimensions were fixed to online specifications.
In game.cpp, created a computeMatrix function that takes in the x, y,
and z position on the breadboard along with an angle to output the
necessary transformation matrix.
Slight edit to render to utilize the new computeMatrix function.
Minor changes to camera position.